+++ RevitAddin.CommandLoader.Tests/Tests.cs
@@ -1,0 +1,12 @@
+using NUnit.Framework;
+
+namespace RevitAddin.CommandLoader.Tests
+{
+    public class Tests
+    {
+        [SetUp]
+        public void Setup()
+        {
+        }
+    }
+}